refactor(backend): post 모듈에서 shared 도메인 분리
FEED 도메인 추가를 위한 사전 작업. recruitment/tips/feed가 공유할
Post, PostImage entity를 post/shared/domain/으로 분리하고,
외부 도메인의 import 경로를 새 위치로 정리.

- post/shared/domain/{post,post-image}.entity.ts 신설
- 기존 post/domain/{post,post-image}.entity.ts 제거
- 외부 import 11곳 경로 업데이트 (chat, comment, member, scrap, admin, seed)
- common/db/schema/index.ts에 shared/domain export 추가

동작 변화 0. 단순 파일 위치 이동 + import 경로 정리.
