+++ apps/backend/src/common/db/seed/main.ts
@@ -3,7 +3,8 @@ import { Pool } from 'pg';
 import { Member, Profile } from '@/member/domain';
 import { Report, ReportCount } from '@/report/domain';
 import { Warn } from '@/admin/domain';
-import { Post, RecruitmentDetail } from '@/post/domain';
+import { Post } from '@/post/shared/domain';
+import { RecruitmentDetail } from '@/post/domain';
 import { Scrap } from '@/scrap/domain';
 import { Participation } from '@/participation/domain';
 import { ChatRoom, ChatRoomMember, ChatMessage } from '@/chat/domain';
