+++ apps/backend/src/member/repository/member.repository.ts
@@ -6,7 +6,8 @@ import { Profile } from '@/member/domain';
 import { DATABASE_CONNECTION, OAuthProvider, Role, Team } from '@/common';
 import { Report, ReportCount } from '@/report/domain';
 import { Warn } from '@/admin/domain';
-import { Post, RecruitmentDetail } from '@/post/domain';
+import { Post } from '@/post/shared/domain';
+import { RecruitmentDetail } from '@/post/domain';
 import { Scrap } from '@/scrap/domain';
 import { Participation } from '@/participation/domain';
 import { RecruitmentQueryResult } from '@/member/type';
