+++ apps/backend/src/post/domain/index.ts
@@ -1,4 +1,2 @@
-export * from '@/post/domain/post.entity';
 export * from '@/post/domain/recruitment-detail.entity';
 export * from '@/post/domain/tips-detail.entity';
-export * from '@/post/domain/post-image.entity';
