+++ apps/backend/src/post/domain/tips-detail.entity.ts
@@ -1,6 +1,6 @@
 import { integer, pgTable } from 'drizzle-orm/pg-core';
 import { baseColumns } from '@/common';
-import { Post } from '@/post/domain/post.entity';
+import { Post } from '@/post/shared/domain/post.entity';
 import { relations } from 'drizzle-orm';
 import { stadiumPgEnum } from '@/common';
