+++ apps/backend/src/post/shared/domain/index.ts
@@ -1,0 +1,2 @@
+export * from '@/post/shared/domain/post.entity';
+export * from '@/post/shared/domain/post-image.entity';
